(PC-17881)[API] refactor: move `young_status` to his own file
