+++ api/tests/core/users/young_status_test.py
@@ -8,7 +8,7 @@ from pcapi.core.fraud import factories as fraud_factories
 from pcapi.core.fraud import models as fraud_models
 from pcapi.core.users import constants
 from pcapi.core.users import factories as users_factories
-from pcapi.core.users import models
+from pcapi.core.users import young_status
 
 
 pytestmark = pytest.mark.usefixtures("db_session")
@@ -21,7 +21,7 @@ class UserStatusTest:
             user = users_factories.UserFactory(
                 dateOfBirth=datetime.datetime.utcnow() - relativedelta(years=age),
             )
-            assert models.young_status(user) == models.Eligible()
+            assert young_status.young_status(user) == young_status.Eligible()
 
         def test_eligible_when_19yo_with_pending_dms_application(self):
             user = users_factories.UserFactory(
@@ -34,39 +34,39 @@ class UserStatusTest:
                     registration_datetime=(datetime.datetime.utcnow() - relativedelta(years=1))
                 ),
             )
-            assert models.young_status(user) == models.Eligible()
+            assert young_status.young_status(user) == young_status.Eligible()
 
     def test_non_eligible_when_too_young(self):
         user = users_factories.UserFactory(
             dateOfBirth=datetime.datetime.utcnow() - relativedelta(years=15) + relativedelta(days=1),
         )
-        assert models.young_status(user) == models.NonEligible()
+        assert young_status.young_status(user) == young_status.NonEligible()
 
     def test_non_eligible_when_too_old(self):
         user = users_factories.UserFactory(
             dateOfBirth=datetime.datetime.utcnow() - relativedelta(years=19, days=1),
         )
-        assert models.young_status(user) == models.NonEligible()
+        assert young_status.young_status(user) == young_status.NonEligible()
 
     def test_beneficiary_when_18yo_and_have_deposit(self):
         user = users_factories.BeneficiaryGrant18Factory()
-        assert models.young_status(user) == models.Beneficiary()
+        assert young_status.young_status(user) == young_status.Beneficiary()
 
     def test_beneficiary_when_underage_and_have_deposit(self):
         user = users_factories.UnderageBeneficiaryFactory()
-        assert models.young_status(user) == models.Beneficiary()
+        assert young_status.young_status(user) == young_status.Beneficiary()
 
     def test_ex_beneficiary_when_beneficiary_have_his_deposit_expired(self):
         user = users_factories.BeneficiaryGrant18Factory(
             deposit__expirationDate=datetime.datetime.utcnow() - relativedelta(days=1)
         )
-        assert models.young_status(user) == models.ExBeneficiary()
+        assert young_status.young_status(user) == young_status.ExBeneficiary()
 
     def test_suspended_when_account_is_not_active(self):
         user = users_factories.BeneficiaryGrant18Factory(isActive=False)
-        assert models.young_status(user) == models.Suspended()
+        assert young_status.young_status(user) == young_status.Suspended()
 
     def test_can_not_mutate(self):
-        status = models.Suspended()
+        status = young_status.Suspended()
         with pytest.raises(attrs.exceptions.FrozenInstanceError):
             status.status_type = constants.YoungStatusType.BENEFICIARY
